Fix default state of system prompt (pull from DB)
